Basic-PDF-Converter-Auto-Contrast-Auto-Brightness-GrayScale-Computer-Vision/requirements.txt
PyMuPDF==1.23.26
Pillow==10.2.0